start working on xh3irq handler
