+++ cpu/rp2350_riscv/cpu.c
@@ -35,29 +35,6 @@ void gpio_reset(void)
     reset_component(RESET_IO_BANK0, RESET_IO_BANK0);
 }
 
-void enable_irq(uint32_t irq_no)
-{
-    uint32_t index = irq_no / 16;
-    uint32_t mask = 1u << (irq_no % 16);
-
-    __asm__ volatile(
-        "csrs 0xbe0, %0\n"
-        : : "r"(index | (mask << 16))
-
-    );
-}
-
-void force_interrupt(uint32_t irq_no)
-{
-    uint32_t index = irq_no / 16;
-    uint32_t mask = 1u << (irq_no % 16);
-
-    __asm__ volatile(
-        "csrs 0xbe2, %0\n"
-        : : "r"(index | (mask << 16))
-    );
-}
-
 /**
  * @brief Initialize the CPU, set IRQ priorities, clocks, peripheral
  */
@@ -89,19 +66,7 @@ void cpu_init(void)
     xosc_sleep(1000);
     printf("Enabling IRQ 50\n");
 
-    enable_irq(51);
-    //
-    // printf("Interrupt force\n");
-
     printf("ra register at start of cpu_init: 0x%lx\n", ra_register);
-
-    printf("Send non-enabled interrupt 50\n");
-    force_interrupt(50);
-    printf("Send enabled interrupt 51\n");
-    force_interrupt(51);
-    printf("Debugger halt message haha\n");
-
-    printf("Back from interrupt :D\n");
 
     __asm__ volatile ("nop");
 }
